Extract inputWithTimeout to jenkinsHelper
