+++ src/org/typo3/chefci/v2/stages/Publish.groovy
@@ -1,8 +1,6 @@
 package org.typo3.chefci.v2.stages
 
 import hudson.model.ChoiceParameterDefinition
-import org.jenkinsci.plugins.workflow.steps.FlowInterruptedException
-import org.jenkinsci.plugins.workflow.support.steps.input.Rejection
 import org.typo3.chefci.helpers.JenkinsHelper
 
 class Publish extends AbstractStage {
@@ -22,49 +20,6 @@ class Publish extends AbstractStage {
     }
 
     /**
-     * Generates a pipeline {@code input} step that times out after a specified amount of time.
-     *
-     * The options for the timeout are supplied via {@code timeoutOptions}.
-     * The options for the input dialog are supplied via {@code inputOptions}.
-     *
-     * The returned Map contains the following keys:
-     *
-     * - proceed: true, if the Proceed button was clicked, false if aborted manually aborted or timed out
-     * - reason: 'user', if user hit Proceed or Abort; 'timeout' if input dialog timed out
-     * - submitter: name of the user that submitted or canceled the dialog
-     * - additional keys for every parameter submitted via {@code inputOptions.parameters}
-     *
-     * @param args Map containing inputOptions and timoutOptions, both passed to respective steps
-     * @return Map containing above specified keys response/reason/submitter and those for parameters
-     */
-    protected Map inputWithTimeout(Map args) {
-        // see https://go.cloudbees.com/docs/support-kb-articles/CloudBees-Jenkins-Enterprise/Pipeline---How-to-add-an-input-step,-with-timeout,-that-continues-if-timeout-is-reached,-using-a-default-value.html
-        try {
-            script.timeout(args.timeoutOptions) {
-                def inputOptions = args.inputOptions
-                inputOptions.submitterParameter = "submitter"
-
-                // as we ask for the submitter, we get a Map back instead of a string
-                // besides the parameter supplied using args.inputOptions, this will include "submitter"
-                Map responseValues = script.input args.inputOptions
-                script.echo "Submitted by ${responseValues.submitter}"
-
-                return [proceed: true, reason: 'user'] + responseValues
-            }
-        } catch (FlowInterruptedException err) { // error means we reached timeout
-            // err.getCauses() returns [org.jenkinsci.plugins.workflow.support.steps.input.Rejection]
-            Rejection rejection = err.getCauses().first()
-
-            if ('SYSTEM' == rejection.getUser().toString()) { // user == SYSTEM means timeout.
-                return [proceed: false, reason: 'timeout']
-            } else { // explicitly aborted
-                script.echo rejection.getShortDescription()
-                return [proceed: false, reason: 'user', submitter: rejection.getUser().toString()]
-            }
-        }
-    }
-
-    /**
      * Asks the user whether the cookbook should be uploaded
      */
     protected publish() {
@@ -74,7 +29,7 @@ class Publish extends AbstractStage {
         def newVersion
         // if we don't get any user response, we do nothing
         if (userInput.proceed) {
-            script.node{
+            script.node {
                 newVersion = bumpVersion(userInput.versionBump)
                 upload()
             }
@@ -106,7 +61,7 @@ class Publish extends AbstractStage {
         def timeoutOptions = [time: 15, unit: 'SECONDS']
 
         // call the input dialog
-        Map inputValues = inputWithTimeout([inputOptions: inputOptions, timeoutOptions: timeoutOptions])
+        Map inputValues = jenkinsHelper.HinputWithTimeout([inputOptions: inputOptions, timeoutOptions: timeoutOptions])
 
         script.echo "Got input ${inputValues}"
 
@@ -131,7 +86,6 @@ class Publish extends AbstractStage {
 
         newVersion
     }
-
 
     /**
      * Uploads the cookbook
